requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docopt==0.6.2
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.2.0
google-auth==2.38.0
google-genai==1.2.0
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pipreqs==0.4.13
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
sniffio==1.3.1
SQLAlchemy==2.0.37
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
WTForms==3.2.1
yarg==0.1.10